debuginfo: extract create.rs
